+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/Navigator.kt
@@ -6,6 +6,7 @@ import androidx.navigation.NavHostController
 import androidx.navigation.NavOptionsBuilder
 import androidx.navigation.serialization.generateHashCode
 import co.electriccoin.zcash.ui.common.provider.ApplicationStateProvider
+import co.electriccoin.zcash.ui.common.provider.IsOledThemeEnabledStorageProvider
 import co.electriccoin.zcash.ui.design.KeyboardManager
 import co.electriccoin.zcash.ui.screen.ExternalUrl
 import co.electriccoin.zcash.ui.screen.about.util.WebBrowserUtil
@@ -27,6 +28,7 @@ class NavigatorImpl(
     private val flexaViewModel: FlexaViewModel,
     private val keyboardManager: KeyboardManager,
     private val applicationStateProvider: ApplicationStateProvider,
+    private val isOledThemeEnabledStorageProvider: IsOledThemeEnabledStorageProvider,
 ) : Navigator {
     override suspend fun executeCommand(command: NavigationCommand) {
         keyboardManager.close()
@@ -67,7 +69,7 @@ class NavigatorImpl(
         }
     }
 
-    private fun replaceAll(command: NavigationCommand.ReplaceAll) {
+    private suspend fun replaceAll(command: NavigationCommand.ReplaceAll) {
         command.routes.forEachIndexed { index, route ->
             when (route) {
                 co.electriccoin.zcash.ui.screen.flexa.Flexa -> {
@@ -102,7 +104,7 @@ class NavigatorImpl(
                     }
 
                     applicationStateProvider.onThirdPartyUiShown()
-                    WebBrowserUtil.startActivity(activity, route.url)
+                    startWebBrowser(route)
                 }
 
                 else -> {
@@ -120,7 +122,7 @@ class NavigatorImpl(
         }
     }
 
-    private fun replace(command: NavigationCommand.Replace) {
+    private suspend fun replace(command: NavigationCommand.Replace) {
         command.routes.forEachIndexed { index, route ->
             when (route) {
                 co.electriccoin.zcash.ui.screen.flexa.Flexa -> {
@@ -145,7 +147,7 @@ class NavigatorImpl(
                     }
 
                     applicationStateProvider.onThirdPartyUiShown()
-                    WebBrowserUtil.startActivity(activity, route.url)
+                    startWebBrowser(route)
                 }
 
                 else -> {
@@ -161,11 +163,11 @@ class NavigatorImpl(
         }
     }
 
-    private fun forward(command: NavigationCommand.Forward) {
+    private suspend fun forward(command: NavigationCommand.Forward) {
         command.routes.forEach { route ->
             when (route) {
                 co.electriccoin.zcash.ui.screen.flexa.Flexa -> createFlexaFlow(flexaViewModel)
-                is ExternalUrl -> WebBrowserUtil.startActivity(activity, route.url)
+                is ExternalUrl -> startWebBrowser(route)
                 else -> navController.executeNavigation(route = route)
             }
         }
@@ -173,6 +175,14 @@ class NavigatorImpl(
         if (command.routes.lastOrNull() in listOf(ExternalUrl, co.electriccoin.zcash.ui.screen.flexa.Flexa)) {
             applicationStateProvider.onThirdPartyUiShown()
         }
+    }
+
+    private suspend fun startWebBrowser(route: ExternalUrl) {
+        WebBrowserUtil.startActivity(
+            activity = activity,
+            url = route.url,
+            isOledDark = isOledThemeEnabledStorageProvider.get() == true
+        )
     }
 
     private fun NavHostController.executeNavigation(