+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/RootNavGraph.kt
@@ -9,6 +9,7 @@ import androidx.navigation.compose.NavHost
 import co.electriccoin.zcash.ui.common.compose.LocalActivity
 import co.electriccoin.zcash.ui.common.migration.MigrationAppHooks
 import co.electriccoin.zcash.ui.common.provider.ApplicationStateProvider
+import co.electriccoin.zcash.ui.common.provider.IsOledThemeEnabledStorageProvider
 import co.electriccoin.zcash.ui.common.viewmodel.SecretState
 import co.electriccoin.zcash.ui.common.viewmodel.WalletViewModel
 import co.electriccoin.zcash.ui.design.LocalKeyboardManager
@@ -33,6 +34,7 @@ fun RootNavGraph(
     val flexaViewModel = koinViewModel<FlexaViewModel>()
     val navigationRouter = koinInject<NavigationRouter>()
     val applicationStateProvider = koinInject<ApplicationStateProvider>()
+    val isOledThemeEnabledStorageProvider = koinInject<IsOledThemeEnabledStorageProvider>()
     val migrationAppHooks = koinInject<MigrationAppHooks>()
     val navController = LocalNavController.current
     val activity = LocalActivity.current
@@ -42,14 +44,16 @@ fun RootNavGraph(
             navController,
             flexaViewModel,
             keyboardManager,
-            applicationStateProvider
+            applicationStateProvider,
+            isOledThemeEnabledStorageProvider
         ) {
             NavigatorImpl(
                 activity = activity,
                 navController = navController,
                 flexaViewModel = flexaViewModel,
                 keyboardManager = keyboardManager,
-                applicationStateProvider = applicationStateProvider
+                applicationStateProvider = applicationStateProvider,
+                isOledThemeEnabledStorageProvider = isOledThemeEnabledStorageProvider
             )
         }